fixed routing architecture to MVC
